+++ client/components/auth-form.js
@@ -2,9 +2,10 @@ import React from 'react'
 import {connect} from 'react-redux'
 import PropTypes from 'prop-types'
 import {auth} from '../store'
-import {BarChart} from './barChart'
-import {LineChart} from './lineChart'
+import {BarChart} from './chartBar'
+import {LineChart} from './chartLine'
 import {MapNeighborhood} from './mapNeighborhood'
+import {CityMap} from './mapCity'
 
 /**
  * COMPONENT
@@ -35,7 +36,7 @@ const AuthForm = props => {
       <a href="/auth/google">{displayName} with Google</a>
       {/* <BarChart /> */}
       {/* <LineChart /> */}
-      <MapNeighborhood />
+      <CityMap />
     </div>
   )
 }
