+++ client/components/chartBar.js
