+++ public/sandbox/SW.json
@@ -1,0 +1,148 @@
+[
+  [-73.710501307789, 40.7272353568104],
+  [-73.7105048727844, 40.7272866599474],
+  [-73.7105564193722, 40.7280291191439],
+  [-73.7106200733055, 40.7288481433292],
+  [-73.7106565417539, 40.729391235726],
+  [-73.7106641135373, 40.7295039801507],
+  [-73.7107535272358, 40.7305336448965],
+  [-73.7108946044347, 40.7309793130038],
+  [-73.7114794046494, 40.7325096721971],
+  [-73.711863270826, 40.7331651083154],
+  [-73.7120822155669, 40.7335267968218],
+  [-73.7122875751903, 40.7338606890303],
+  [-73.7125023650794, 40.7342099041115],
+  [-73.7127890454674, 40.734678652076],
+  [-73.7128314007636, 40.7347478981865],
+  [-73.7128996067261, 40.7348594076669],
+  [-73.7132879772167, 40.7354813959322],
+  [-73.7137029138319, 40.7361683938415],
+  [-73.7143676492832, 40.7372521635011],
+  [-73.7147711839485, 40.7379086196946],
+  [-73.7149419872024, 40.7382428715638],
+  [-73.7149633553637, 40.7382846821137],
+  [-73.7151125691164, 40.7385766813613],
+  [-73.7154508248542, 40.7392250760218],
+  [-73.7157845597965, 40.7399026179418],
+  [-73.7165690126453, 40.741657731049],
+  [-73.7174073429879, 40.7435192029459],
+  [-73.7176044191195, 40.7439829014029],
+  [-73.7176265212034, 40.7440367521552],
+  [-73.7176443877215, 40.7440786868556],
+  [-73.7182338601329, 40.7454049359345],
+  [-73.7185319950384, 40.7460476034137],
+  [-73.7186917323382, 40.7464010391076],
+  [-73.7189227066734, 40.7469031418173],
+  [-73.7189454216533, 40.74695252815],
+  [-73.7189650416389, 40.7470030136543],
+  [-73.7198879996326, 40.7493777832946],
+  [-73.7201023738974, 40.7497836027211],
+  [-73.7201456534227, 40.749771104333],
+  [-73.7225870779975, 40.7488587251441],
+  [-73.7233973226397, 40.7503973492807],
+  [-73.7262734938982, 40.7493147644061],
+  [-73.7276816882985, 40.7496475180395],
+  [-73.7277662723552, 40.7494649890817],
+  [-73.7278420931925, 40.7492802466095],
+  [-73.7279090480517, 40.7490935335347],
+  [-73.7279670495635, 40.7489050946052],
+  [-73.7280160233643, 40.7487151800005],
+  [-73.7280559022018, 40.7485240390146],
+  [-73.7282919212032, 40.7474669380461],
+  [-73.7288875772188, 40.7475749098871],
+  [-73.7289631922887, 40.7476411421101],
+  [-73.7287208899652, 40.7482053540052],
+  [-73.7286234358707, 40.7485203653105],
+  [-73.7284931472538, 40.7490529804588],
+  [-73.728379345704, 40.7493894400725],
+  [-73.7282551718074, 40.749702001656],
+  [-73.7280549181677, 40.7500904244058],
+  [-73.7278853730732, 40.750395756302],
+  [-73.7277149235766, 40.7506351684391],
+  [-73.7273802532006, 40.7510378406149],
+  [-73.727095257514, 40.7513498787756],
+  [-73.7267787560691, 40.7516313655269],
+  [-73.7266666611558, 40.7517218620223],
+  [-73.7264160830994, 40.7519241512163],
+  [-73.7260310243125, 40.7521826289087],
+  [-73.7258312500637, 40.7522999943598],
+  [-73.7250958144748, 40.7527158648951],
+  [-73.7250698415124, 40.7527301174174],
+  [-73.7247595916586, 40.7529164223887],
+  [-73.7238458788836, 40.7534650856996],
+  [-73.7238283495944, 40.753477363554],
+  [-73.7235200330738, 40.7537297114317],
+  [-73.723442715369, 40.7537991190549],
+  [-73.7232680953145, 40.7539558738261],
+  [-73.7229943725217, 40.7542259461844],
+  [-73.722798391402, 40.7544418123822],
+  [-73.7226467959872, 40.7546087982732],
+  [-73.7224123081378, 40.7549284629304],
+  [-73.7220720745542, 40.7555049775227],
+  [-73.7218042520997, 40.7560781728924],
+  [-73.7217647561916, 40.7561467422836],
+  [-73.7216149309818, 40.7564068699013],
+  [-73.7215976744181, 40.7564357795895],
+  [-73.7214600624668, 40.7566663556664],
+  [-73.7212131210706, 40.7569874843531],
+  [-73.7211113741724, 40.7571198012183],
+  [-73.7209664771016, 40.757302063814],
+  [-73.720634553224, 40.7576251105694],
+  [-73.7203173121871, 40.7579055330636],
+  [-73.7199843295208, 40.7581690966973],
+  [-73.7196786821524, 40.7583807745814],
+  [-73.7195659672165, 40.758458835587],
+  [-73.7191804551269, 40.7586747130906],
+  [-73.7189718620393, 40.7587915129124],
+  [-73.7187792357878, 40.7588799672712],
+  [-73.7183210773124, 40.7590903515289],
+  [-73.7176737299849, 40.7593194873466],
+  [-73.716855762659, 40.75953379035],
+  [-73.7161397846841, 40.7596665546181],
+  [-73.7161266406061, 40.7596689896264],
+  [-73.7149834977723, 40.7597911407247],
+  [-73.7145582174635, 40.7598170841521],
+  [-73.7142642350896, 40.7598350229289],
+  [-73.7135196389212, 40.7598460068673],
+  [-73.7133267788453, 40.7597384655125],
+  [-73.7124507631443, 40.7592499799875],
+  [-73.7108590458291, 40.7582033405671],
+  [-73.7105463541221, 40.7579993102044],
+  [-73.707826741321, 40.7562246800826],
+  [-73.70527263831, 40.7547787166841],
+  [-73.7034691734821, 40.7536916383749],
+  [-73.7027363098763, 40.7532587919835],
+  [-73.7021797707117, 40.7528942152767],
+  [-73.7016952230202, 40.7525451300912],
+  [-73.7016348128049, 40.7525016041195],
+  [-73.7016234750995, 40.7524337923847],
+  [-73.701348512789, 40.7507888562835],
+  [-73.7011745185119, 40.7489325752952],
+  [-73.7009026257981, 40.7471368175869],
+  [-73.7008947532951, 40.7470655396473],
+  [-73.700780960888, 40.7450497273468],
+  [-73.7005829920485, 40.7431921145571],
+  [-73.7002935737127, 40.7410188746583],
+  [-73.7000227564256, 40.7394124384208],
+  [-73.7000104153247, 40.7393316852096],
+  [-73.7000215564582, 40.7392448138448],
+  [-73.7003197043123, 40.7389763547386],
+  [-73.7008128232806, 40.7384474248056],
+  [-73.7014720822183, 40.7375198324041],
+  [-73.702061716163, 40.7364510284972],
+  [-73.7022821947228, 40.7361340678727],
+  [-73.7026892982351, 40.7354984946449],
+  [-73.7033162619515, 40.7345715284755],
+  [-73.7035238436534, 40.7342162162545],
+  [-73.7039459842085, 40.7335345330818],
+  [-73.7046024971219, 40.7325780178265],
+  [-73.7049967160781, 40.7319788289609],
+  [-73.7052230113085, 40.7316279011076],
+  [-73.7058487004352, 40.7306811251852],
+  [-73.7064816352596, 40.7297195697141],
+  [-73.7071010555002, 40.728755984371],
+  [-73.7076635263988, 40.7278392038416],
+  [-73.7085708445439, 40.7276446883622],
+  [-73.7104324500146, 40.7272493307414],
+  [-73.710501307789, 40.7272353568104]
+]
